cleaned main code
