+++ src/lib.rs
@@ -6,9 +6,14 @@ pub mod gui;
 pub mod model;
 pub mod parsing;
 pub mod render;
+pub mod events;
+
+pub const SCREEN_WIDTH: u32 = 1600;
+pub const SCREEN_HEIGHT: u32 = 900;
 
 pub fn run() {
     let scene = getScene();
+    event_loop(&scene);
 }
 
 pub fn event_loop(scene: &Scene) {
